setup for v3 meshing, working on setting up pointcloud winding datastructure

+++ thaumato-anakalyptor/ThaumatoAnakalyptor/pointcloud_mesh_refinement.py
@@ -50,7 +50,7 @@ def generate_winding_pointclouds(mesh_path):
     slabs = glob.glob(os.path.join(base_path, "points_selected_", "*.npz"))
     print("Preparing the winding pointclouds...")
     for slab_path in tqdm(slabs, desc="Preprocessing slabs"):
-        points = load_pointcloud_slab(slab_path)
+        points = load_pointcloud_slab(slab_path).astype(np.float16)
         points = np.concatenate((points, points_mesh), axis=0)
         min_angle = np.min(points[:, 3])
         max_angle = np.max(points[:, 3])
@@ -66,7 +66,7 @@ def main():
     parser = argparse.ArgumentParser(description="Refine pointcloud using mesh and downsampling")
     parser.add_argument("--mesh", required=True, help="Path to mesh file (.ply or .obj)")
     args = parser.parse_args()
-
+    
     generate_winding_pointclouds(args.mesh)
